Use git clone instead of pip
Use the version that is in the git repo instead of what is pushed to
pypi. This gives a more recent version in the image.

Remove the default main as there now is an index page that will be shown
if you don't specify something. Also remove the path in the entrypoint
and rely on '.' to be used (which is set to /data inte the workdir
option).